\name{PracticaAYM-package}
\alias{PracticaAYM-package}
\alias{PracticaAYM}
\docType{package}
\title{
\packageTitle{PracticaAYM}
}
\description{
\packageDescription{PracticaAYM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PracticaAYM}
\packageIndices{PracticaAYM}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{PracticaAYM}

Maintainer: \packageMaintainer{PracticaAYM}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
##suma.dos.numeros(x,y)
}
